Dispense drafted
There are probably some features missing, but the majority of the code
for the dispense function has at least been drafted in. Ther are some
constants that need to be updated for the actual device, and line 191
needs to be updated to properly show an error message.

The dispense function also needs to have the "success" portion of the
code added, as the current function only deals with error checking,
mainly.